se crearon los services y el endpoint de listar y crear usuarios

+++ src/main/java/com/backend/apartamentos/dtos/users/response/PublicUserDTO.java
@@ -1,6 +1,6 @@
 package com.backend.apartamentos.dtos.users.response;
 
 public record PublicUserDTO(
-        Long id, String name, int age
+        int id, String name, int age
 ) {
 }
